LayoutTests/fast/dom/timer-unthrottle-on-scroll-expected.txt

Tests that a repeating timer changing the style of an element outside viewport gets unthrottled when scrolling the element into view.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isTimerThrottled(timeoutId) is false
5th iteration, timer should be throttled as the element is outside the viewport.
PASS wasThrottled is true
Timer should still be throttled before scrolling.
PASS internals.isTimerThrottled(timeoutId) is true
Scrolling element into view.
Timer should no longer be throttled
PASS internals.isTimerThrottled(timeoutId) is false
6th iteration, timer should still be unthrottled.
PASS internals.isTimerThrottled(timeoutId) is false
PASS successfullyParsed is true

TEST COMPLETE
